doc: add a "developing with dune" tutorial
This adds a new tutorial (our first one!) about common interactions with
dune, how to structure a project, using diffing and promotion, etc. The
goal is to cover the 20% simple parts of dune that will be used 80% of
the time.

I relied heavily on https://diataxis.fr/ for that, in particular I don't
consider that this is finished work, just that it's now just complete
enough to be integrated in main. I did some user testing on the first
half, and I consider that the first half is pretty OK, while the second
part (starting from C stubs) will need more work to be usable, but I'll
continue improving that in main with more user testing.

A technical aspect of that PR is that I decided to go with myst
(markdown parser for sphinx) - see the syntax guide. I don't want to
make it a policy, but I think for new documents it makes sense to use
something like that, because:

- it maps fully to reST
- it's more contributor friendly
- editor tooling is better
- it alleviates some vendor lock-in (since we're likely to migrate to
  ocaml.org at some point, and it's more likely to use a variant of md
  than sphinx support odoc syntax)

The page is marked as orphan and WIP for now since it's not meant for
end users yet.

Signed-off-by: Etienne Millon <me@emillon.org>