Fix rule target conflict with extra source (#10706)
* Fix rule target conflict with extra source

This is a quick fix that addresses the "Error: Multiple rules
generated for <target>" when building a package with the extra-source
field. This isn't expected to work in cases where a package has
multiple extra-source entries, which will come in a later change.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>

* Update test output after rebasing

Signed-off-by: Marek Kubica <marek@tarides.com>

* Fix test so it does not accidentally vendor `needs-patch`

Signed-off-by: Marek Kubica <marek@tarides.com>

* Remove git dir from extra-sources test tarball

Workaround for an issue where our oneshot webserver disconnects from
the client too early when sending large files. Remoning the .git
directory from the tarball used in the extra-sources test reduces the
size of the tarball to avoid the problem.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>

---------

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>
Signed-off-by: Marek Kubica <marek@tarides.com>
Co-authored-by: Marek Kubica <marek@tarides.com>