Make [Pkg.t]s for system-provided packages
The value of the PATH environment variable when building a pacage is
set such that all the executables in that package's dependency cone
are visible. This includes the ocaml compiler itself. In the past,
system-provided packages (such as dune) were omitted from this
dependency cone, along with their transitive dependencies.

This creates a problem for packages that depend on dune and don't
directly depend on the compiler. Such packages typically build with
`dune build -p ...`, and dune assumes that it's being run in an
environment where the ocamlc compiler is visible. Although the dune
package depends on the compiler, due to its exclusion from the
dependency cone used to set the PATH variable dune is unable to find
the compiler when building the package.

We haven't noticed this problem until now because it's rare to use
dune in environments where the compiler is not globally installed or
present in the active opam switch, however running dune on systems
without a global ocaml installation is something we're trying to
enable in the near future.

The fix is to add all packages, system-provided or not, to the graph
of [Pkg.t] objects that is used to compute the build environment for
each package. To prevent dune from attempting to build system-provided
packages, the check whether a package is system-provided is deferred
to the point where dune computes the list of packages to build.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>