Add toolchains feature
When enabled, this causes packages that look like the ocaml compiler
to be installed in a user-global location rather than locally within
the project. This is a workaround for the problem where the compiler
can't be relocated, and allows the compiler binaries to be reused
across different projects without the use of the dune cache.

Compiler packages are built and installed using the build and install
commands from the package's lockfiles. This avoids the need to
download additional compiler package opam files in order to learn how
to build the compiler at build-time, and means that users can modify
the compiler build instructions by changing them directly in the lock
directory. This also allows toolchain packages to be pinned like
regular packages.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>