Implement post dependencies
When computing the dependencies of individual packages, distinguish
between post and regular dependencies. This is done by first
evaluating a package's dependency formula with the same environment as
was used for solving (with post=true) to compute all dependencies,
then computing non-post dependencies by evaluating it a second time
with post=false, ignoring any additional packages added to the formula
or any terms that now have no solutions as a consequence of this
change to the environment.

Note that this change also sets post=true during solving which matches
the behavior of Opam's solver.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>