Consider post dependencies when building packages
This makes the post dependencies of a package available in the build
environment of that package's dependant packages. Note that the post
dependencies of a package can't be present in that package's own build
environment, as its possible for a post dependency of a package to
depend on that package itself, and thus it must be possible to build
the post dependencies of a package after build that package.

However in a situation where a package foo has a post dependency on a
package bar, and a third package baz depends on foo, baz should have
access to both foo and bar in its build environment.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>