Replace oneshot_webserver with fake curl
This solves a lot of the waiting leftover HTTP processes etc. The new
fake curl will read the list of files to serve from `fake-curls` and map
each line to a port.

To match the behavior of the oneshot-webserver it will only serve each
port once, subsequent requests for the same port will be answered by 404
(so not exactly the same but similar enough), the ports that were served
are recorded into an `already-served` file.

The same file can be served multiple times, as long as it is present
multiple times in `fake-curls`.

Signed-off-by: Marek Kubica <marek@tarides.com>