feat(test): Make it possible to disable line wrapping
Sometimes cram tests are not reproducible if different platforms output
different length errors, in which case even editing out the errors with
common tools like `sed` does not help as the wrapping happens before the
`sed` invocation runs, thus the message is split over multiple lines.

This introduces an option to disable the automatic wrapping of lines.

Signed-off-by: Marek Kubica <marek@tarides.com>