Return [result] from [Workspace_root.create]
This makes it possible to check whether the user is currently inside a
dune project. This is needed for dev-tools to have fallback behaviour
when run outside of a dune project which is necessary to allow users
to configure their editors to invoke dev-tools via dune. For example,
if a user configures their editor to start ocamllsp with the command
`dune ocamllsp`, dune needs to be able to handle the case where the
user edits an ocaml file outside of a dune project, for example by
attempting to invoke ocamllsp from PATH.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>