Add command to build and run ocamllsp
This command will exit with an error message unless it's run within a
dune project with a dune.lock directory containing a lockfile for the
package "ocaml" so it's not ready for use by editors. Support for
falling back to a system or opam-managed installation of ocamllsp will
come in a later change.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>