Change colors to be more readable in light themes
Some of the text colors used by dune tend not to be readable on
default light-themes terminals. Of course it will always be possible
to configure your terminal so that some colors are too light/dark to
read, but some of the preset themes on macos are unable to display
some colors commonly used by dune.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>