Fix bug updating lockdir inside subdirectory
To atomically update a lockdir we rename the existing lockdir to begin
with a ".". Prior to this change the entire relative path to the
lockdir was renamed which meant that if the lockdir was inside a
subdirectory (e.g. a dev tool lockdir) then dune would attempt to
rename "path/to/lockdir" to ".path/to/lockdir". This fails because
rename/mv doesn't create directories to satisfy the destination
path. This change fixes the issue by instead renaming the lockdir to
"path/to/.lockdir" which avoids the need to create any directories.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>