Configurator: try to call pkgconf before failing back to pkg-config
Pkgconf is a modern implementation of the "pkg-config freedesktop standard". Linux distros actually use pkgconf as their
pkg-config implementation and provide a in place replacement. (Homebrew packages both and is the last place I found where you can get the historical implementation).
It is always OK to call pkgconf for pkg-config all the CLI and outputs within the standard are following the standard :-).

Cygwin pkg-config package though provides pkgconf but does not provide the alias pkg-config forcing us to uptream the switch to the tool name instead of the standard.

Signed-off-by: Pierre Boutillier <pierre.boutillier@laposte.net>