Fix the compilation of OCaml 5 programs on Cygwin
strdup is not part of the C standard library but is part of POSIX.
By default GCC and Clang both use the GNU variant of the C standard,
so we are using strdup here to ensure users have access to it by default.
If -std=c++11 is used instead of -std=gnu++11 this would fail to compile
on non-POSIX platforms such as Cygwin.

Signed-off-by: Kate <kit-ty-kate@outlook.com>