pkg: relax version constraints for ocamlformat dev tool
Previously dune would install the exact version of ocamlformat specified
in the .ocamlformat file in the project's root. This prevents
alternative distributions of the ocamlformat package from being used, as
these typically append a suffix to the version number.

The motivation for this change is supporting a binary dev tools
distribution where the ocamlformat package is named like
"ocamlformat.0.26.2+binary", and we want that package to be installable
when the .ocamlformat file of a project specifies "version=0.26.2".

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>