feature: add aliases for inline tests
For a given library with inlines tests, we add an alias with the name of
the library for running those tests.

Further work on separating out the partition actions would allow us to
have an alias for each partition.

For now, this allows us to run inline tests with the runtest command.

Signed-off-by: Ali Caglayan <alizter@gmail.com>

<!-- ps-id: 8fef343c-46e6-4486-aa50-d31b6cfe1d46 -->

Signed-off-by: Ali Caglayan <alizter@gmail.com>