pkg: ocamlformat dev tool requires lockdir
This is a workaround for an issue where dune needs to locate an ocaml
toolchain in order to run `dune fmt`, and in the absence of a lockdir
dune will look in PATH for ocaml tools, which may not be present. The
consequence of this issue is that users of the developer preview can
have an ocaml toolchain installed by dune (in their
~/.cache/dune/toolchains directory), and running `dune fmt` will
complain about `ocamlc` not being found. Forcing the lockdir to exist
means that dune will always take the toolchain implied by the lockdir,
and there won't be a situation where the command fails due to `ocamlc`
being missing.

More info in this issue:
https://github.com/ocaml/dune/issues/11038

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>