Configurator: pkg-config plugin uses pkgconf and --personality=TARGET by default (#10937)
* configurator: Allow to specify a path in PKG_CONFIG

This makes the implementation of `which` in configurator closer to the one of stdune.

Signed-off-by: Pierre Boutillier <pierre.boutillier@laposte.net>

* Configurator: try to call pkgconf before failing back to pkg-config

Pkgconf is a modern implementation of the "pkg-config freedesktop standard". Linux distros actually use pkgconf as their
pkg-config implementation and provide a in place replacement. (Homebrew packages both and is the last place I found where you can get the historical implementation).
It is always OK to call pkgconf for pkg-config all the CLI and outputs within the standard are following the standard :-).

Cygwin pkg-config package though provides pkgconf but does not provide the alias pkg-config forcing us to uptream the switch to the tool name instead of the standard.

Signed-off-by: Pierre Boutillier <pierre.boutillier@laposte.net>

* Configurator tests: toy pkg-config cleaned and renamed into pkgconf

Signed-off-by: Pierre Boutillier <pierre.boutillier@laposte.net>

* Configurator: add by default --personality=@target@ to pkgconf calls

Specifying a personality is only useful when cross-compiling (which you do when you use mingw on Windows) but
(nearly) never harmful as pkgconf installs a default.personality which it uses for all triplets (that parses
therefore the nearly) it hasn't a explicit personality for.

Signed-off-by: Pierre Boutillier <pierre.boutillier@laposte.net>

* Configurator: document new pkg-config plugin behaviors

Signed-off-by: Pierre Boutillier <pierre.boutillier@laposte.net>

---------

Signed-off-by: Pierre Boutillier <pierre.boutillier@laposte.net>