feature: improve hints for dune runtest
We improve the hints given by `dune runtest` by letting it also hint
about directories, which was missing before. This allows directory cram
tests to also be suggested when a user mispells a test name.

We also improve the "not found" message to one about a test not being
found in case the user passes a non-test file.

Signed-off-by: Ali Caglayan <alizter@gmail.com>

<!-- ps-id: 858fb282-90f0-49bb-802c-feb9aa4d0af3 -->