Allow arbitrary opam depopt disjunction represntations
Opam currenly represents lists of depopts packages with disjunctions
that "stack" no the left-side, such as: `Or(Or(A, B), C)`. This code was
originally written to expect disjunctions to stack on the right-side
(`Or(A, Or(B, C))`). This change generalizes the resolving of depopts so
that the representation does not matter, as long as every node is `Or`,
`Atom`, or `Empty`.

Since we now traverse both sides of the `Or` node it's tricky to keep
the `collect` function tail-recursive so the code is simplified to avoid
collecting the atoms into an accumulator. Packages are expected to have
a small number of depopts so there's little benefit to keeping it
tail-recursive given the code complexity it would add.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>