fix: Do not depend on `dune` in `dune-project` in `dune init` template
Dune automatically adds the right constraint anyway, with the right
version constraint. All Dune project implicitely depend on Dune anyway
and the tests show that the opam files correctly include Dune.

This helps with #11106 but also avoids the bug where Dune couldn't solve
a package because it depended on Dune and in the Dune Package Management
we explicitely remove the dependency.

Signed-off-by: Marek Kubica <marek@tarides.com>