pkg: version comparison fastpath
Solving dependencies requires performing many package version
comparisons. Opam's logic for comparing versions is complicated in the
worst case, however in practice most package versions follow an
approximation of semantic versioning. In such cases the components of
the version can be packed into a single integer value which can be
efficiently compared.

This change introduces a pre-processing step for package versions that
packs them into an int of possible and uses int comparison when
comparing package versions when possible.

This is based on an optimization in python's uv package manager.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>

<!-- ps-id: 4a98791d-3953-418e-ad88-08bd8fd24bfa -->

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>