feat(pkg): cache revision store
We introduce a caching mechanism for the revision store. This cuts
solving time by around 50% on average.

It works by using LMDB to cache both the ls-tree tree operations and the
file ontents of an object. These lookups happen frequently when running
the solver, so any time we can save here is quite noticable.

For instance, a locking that takes around 800ms now takes around 400ms
when cached.

LMDB may also be used to cache other lookups in dune in the future, for
instance the parsing of dune files.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>
Co-authored-by: Ali Caglayan <alizter@gmail.com>