refactor(pkg): irrefutable constraints
Reduce loading of unnecessary opam packages by making sure that every
local package must not conflict with the constraints of local packages.
Such packages can never be used for a solution, and loading them is
pointless.

This is a large speed improvement for repos that write many constraints
(such as bonsai for example).

<!-- ps-id: cf770d91-1c12-43a4-aae7-2e721575acc7 -->

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>