pkg: defer evaluating platform vars in commands
Package build and install commands are expressions that generate the
command to run by evaluating a series of conditional statements. This
allows the commands to be tailored to the current machine where the
command is being run. Previously dune would evaluate platform-specific
variables (e.g. os, arch) in these expressions at solve-time, producing
lockfiles specialized to the machine where they are solved. This change
defers evaluation of these variables until build-time.

The consequence of deferring evaluation of platform variables to build
time is that lockfiles will preserve more of the conditional logic found
in the corresponding opam file. Note that some variables are still
evaluated, such as `with-test`, as these variables are not based on the
platform. The benefit of deferring evaluation of platform variables is
that users of computers that differ from the computer where the lockfile
was generated will be able to run the build/install commands specialized
for their computer, which is more likely to work than attempting to run
the build/install commands specialized for the computer where the
lockfile was generated.

Note that this change alone is not sufficient for making lockfiles
portable, as it doesn't address the fact that the dependencies of a
package can also vary depending on platform-specific variables.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>