test: cram tests are run twice when both aliases are built
This bug was noticed when working on #11109. The issue is that the cram
rules create two copies of the anonymous cram action, however since the
alias names are different (runtest vs cram test name) this ends up with
having two different digests. This means that when building `@runtest` and
`@mycram` we end up running the cram test `mycram.t` twice.

This test demonstrates this behaviour.

Signed-off-by: Ali Caglayan <alizter@gmail.com>