fix: double run of cram tests
We fix an issue where cram tests were being run twice due to anonymous
actions being created for each alias associated to a given cram test. We
fix this by distinguishing one of the aliases and only creating the
action for that while allowing the other aliases to depend on that
alias.

This required changing the cram specification slightly, but has
simplified the rule generation somewhat.

Signed-off-by: Ali Caglayan <alizter@gmail.com>