pkg: remove broken symlinks from package sources
Dune cannot handle symlinks with missing targets. Package source
archives may contain such broken symlinks. This can be due to packaging
errors on the part of the package author which are benign to the correct
operation of the package when installed with opam, but render packages
uninstallable with dune.

It's assumed that all opam packages can be installed with opam, so if a
package contains a broken symlink then the symlink must not be necessary
for the correct operation of the package. Thus the fix is to remove
broken symlinks from extracted source archives, and to avoid copying
broken symlinks from local package source directories.

A less conservative approach than this might attempt to allow dune to
handle broken symlinks in general, but this would be a much larger
change, and has been attempted and abandoned several times before.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>