Add the warning flag to the C/C++ compiler invocation (#11535)
* Differentiate the ccomp_type and the cc_vendor

ccomp_type is given by the OCaml compiler. It's either 'cc' or 'msvc',
but could be something else in future versions. The C compiler vendor
is another specialization. This aligns with the OCaml compiler
terminology.

Signed-off-by: Antonin Décimo <antonin@tarides.com>

* Rename Cxx_rules and Cxx_flags to Cc_rules and Cc_flags

They describe how to call the C/C++ compiler and are not directly tied
to C++ anymore.

Signed-off-by: Antonin Décimo <antonin@tarides.com>

* Fix detection of clang-cl in cc_vendor

clang-cl is a drop-in replacement for MSVC, and exposes the same set
of flags as MSVC. It defines both `_MSC_VER` and `__clang__`, but we
want it do have precedence over clang, just as clang (on Unix) defines
both `__GNUC__` and `__clang__`, which is why we want it to have
precedence over GCC.

Signed-off-by: Antonin Décimo <antonin@tarides.com>

* Disable MSVC copyright output when discovering the cc_vendor

MSVC has an annoying ‘feature’: it prints a copyright statement
everytime it's called. Fortunately, it can be silenced by passing the
`/nologo` argument. MSVC always prints the name of the file it's
compiling.

Signed-off-by: Antonin Décimo <antonin@tarides.com>

* Improve warning message in C/C++ compiler detection

Signed-off-by: Antonin Décimo <antonin@tarides.com>

* test: rename some variables for consistency

echo-ing the variables is somewhat brittle but removes the spaces
inbetween flags.

Signed-off-by: Antonin Décimo <antonin@tarides.com>

* Add a warning flag to the C/C++ compiler invocation

Note that as long as we don't set the warning as error flag (-Werror
on cc or -WX for msvc), compilation can't fail. No workflow will be
broken, the user will just see new warnings reported by the compiler.

Signed-off-by: Antonin Décimo <antonin@tarides.com>

---------

Signed-off-by: Antonin Décimo <antonin@tarides.com>