fix: pass pkg-config (extra) args in all pkg-config invocations
pkgconf uses the personality to locate the correct library when the
library can be found with multiple toolchains. For instance, on
Cygwin, pkgconf could detect libcurl (built with cygwin1.dll), libcurl
(built with x86_64-w64-mingw32), or libcurl (built with
i686-w64-mingw32). It needs the right `--personality` flag to select
between these libraries, which we pass when asking for the `--cflags`
or `--libs`. We also need it when validating the expression, or
pkgconf returns an error:

    $ (cd _build/default && config/discover.exe)
    which: pkgconf
    -> found: C:\Users\Antonin\AppData\Local\opam\default-2\bin\pkgconf.exe
    run: C:\Users\Antonin\AppData\Local\opam\default-2\bin\pkgconf.exe --print-errors "libcurl >= 7.28.0"
    -> process exited with code 1
    -> stdout:
    -> stderr:
     | Package libcurl was not found in the pkg-config search path.
     | Perhaps you should add the directory containing `libcurl.pc'
     | to the PKG_CONFIG_PATH environment variable
     | Package 'libcurl' not found
    Error: Package libcurl was not found in the pkg-config search path.
    Perhaps you should add the directory containing `libcurl.pc'
    to the PKG_CONFIG_PATH environment variable
    Package 'libcurl' not found

Signed-off-by: Antonin Décimo <antonin@tarides.com>