Allow concurrent build with RPC server
Allow the `dune build` command to be run concurrently with an RPC
server (in either passive or eager watch mode). This works by having the
second instance of dune send RPC build command to the RPC server, and
the RPC server then sends diagnostic messages back to the first instance
of dune.

Currently this is only implemented for the `dune build` command, and
promoting changes into source files is not supported. There are also
some subtle differences between the output of `dune build` is performing
the build itself and when it's acting as an RPC client, such as file
paths being absolute rather than relative and the underline under code
errors losing its color.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>