refactor(foreign): make Foreign_language.t a poly variant
We make foreign_language.t a polymorphic variant to allow for more fine
grained casing. Later, as outlined in #11849, we will add more foreign
languages, so it is important that we can refer to a subset of the
foreign languages for later language specific rules.

We also do some routine cleanup and dead code removal.

Signed-off-by: Ali Caglayan <alizter@gmail.com>