Remove some unused logic from exec (#11823)
This removes some code related to automatically adding a missing ".exe"
suffix from a relative path to an executable passed as the program
argument to `dune exec ...` if doing so would result in the name of an
executable build target. However the condition is seemingly inverted
from what is should be; the suffix is only added if the suffix already
exists. Since nobody seems to complain about this bug, let's take the
opportunity to delete the code altogether.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>