feat(cram): cleanup subprocesses on exit
This should fix a common problem where users run subprocesses, for
example using `./script.sh &` inside their cram tests. If these
subprocesses do not terminate by themselves, they become orphaned and
have to be manually killed. A common use case is testing a web-server.

We add a more sophisticated trap for the underlying shell script of a
cram test so that it can successfully terminate all of its children.

Signed-off-by: Ali Caglayan <alizter@gmail.com>