General portable depexts in lockfiles
Dune stores depexts for each package in the lockdir. These are the names
of system dependencies that must be installed by the system package
manager (e.g. apt-get, homebrew) in order for the package to be used
(possibly at build time or at run time). The name of a system package
may vary between package managers, so opam package typically contain a
list of different names for a depext, along with conditions under which
that name should be used. For example the conf-pkg-config opam package
specifies that it depends on a depext named "pkg-config" on ubuntu, but
a depext named "pkgconf" on archlinux.

Dune uses depexts to provide hints when a package fails to build. If a
package with at least one depext fails to build, dune will print a
message suggesting to the user that the reason the package failed to
build might be because of a missing depext, and it will try to print the
depext name appropriate to the user's system. This is because the
expectation is that the user will the go and manually invoke their
package manager to install the package. (Unlike opam, dune does not
currently attempt to invoke the user's package manager and install
depexts automatically).

Dune can also be queried for a list of all depexts for the current
project.

Currently dune only stores the names of depexts specialized for the
system where the lockdir was generated. This creates a problem for users
working on a project with a checked-in lockdir on a computer other than
the one where the lockdir was generated. Dune will tell users possibly
incorrect names for their project's depexts, since the names will be
tailored for a different type of computer with a different package
manager where the appropriate depext has a different name.

This commit changes the behaviour of the `depext` field in lockfiles to
capture all of the depext metadata from the corresponding opam file,
converted to dune's slang dsl. This allows the choice of which depext
name to show to be deferred to build time, which means that the depext
names will always be appropriate for the current system.

This new behaviour is behind the "portable-lock-dir" feature flag.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>