Warn about some common typos in pcakage variables (#11832)
Dune uses underscores to separate words in pform variables but package
variables use dashes for compatibility with opam. This will probably
lead to typos. Technically arbitrarily-named solver variables are
allowed by opam so typos in variable names have the potential to
be silently accepted by dune while having no effect on the solution.
This change prints a warning when dune encounters a package variable
name that is identical to a known variable except with dashes replaced
with underscores.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>