fix: dune test dirtest.t/run.t running cram test incorrectly
We fix an issue where a user can write

```
dune test dirtest.t/run.t
```

and the corresponding cram test was being run incorrectly. This was due
to the buggy parent detection code in runtest.ml. We fix how the parent
directory of a cram test is actually detected and do some cleanup for
cram source related code.

The runtest.ml test is updated to better reflect what we expect the
command to do.

Signed-off-by: Ali Caglayan <alizter@gmail.com>