fix(exec): Spawn dune exec processes in inherited process group
In dune 3.18 we invoked Spawn directly to start a process and in turn we
defaulted to inheriting the process group. In 3.19 this was changed to
use the process managment given by Dune_engine.Process however there we
always spawn processes in a new process group.

Terminal sessions typically rely on the process group when doing
interactive things like reading lines from the user. This meant we ran
into issues such as #11870 and #11867.

The fix here is to revert back to inheriting the process group.

Signed-off-by: Ali Caglayan <alizter@gmail.com>