fix: always close fd when creating cram script
When generating a shell script for a cram test we make sure to always
close the file descriptor even if we raise an exception.

For example, on Windows, we might raise an exception due to
translate_path_for_sh when the cygpath.exe binary is not available. This
would cause the main.sh script to be left open and leave dune unable to
cleanup the sandbox for the cram test.

Signed-off-by: Ali Caglayan <alizter@gmail.com>