Looking up paths to executables with dune exec
The "dune exec" command has three different ways of resolving the
names of executables to paths to executables:
 - public executables defined in the current project
 - executables in the "bin" directories of dependencies
 - executables in directories listed in $PATH

This can lead to unexpected shadowing, especially in the case of
executables from dependecies, as users may not be aware that one of
the packages in their project's dependency cone defines an executable
with the same name of an executable that's also insalled system-wide.

Short of fixing the problem for now, this change introduces a tool for
helping investigate specifically which executable will be run by "dune
exec". This adds a flag "--which" to "dune exec" which causes the path
to the executable to be printed, rather than running the executable.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>