feat(pkg): ocaml implementation of dune_patch
We vendor hannesm/patch in order to get a portable implementation of our
patch action. This removes the dependency on an external patch/gpatch
binary, making the patch action work reliably across all platforms.

- Fixes #12232
- Fixes #7699

We also reorganise and extend the test suite for the patch action.
Making sure that we test the more difficult parts of patching with good
coverage.

Signed-off-by: Ali Caglayan <alizter@gmail.com>