[rocq] Initial support for the Rocq Prover build language
We introduce `(lang rocq)`, a build mode for the Rocq Prover,
previously known as "Coq".

The build mode is an exact copy of the `(lang coq)` mode, renaming the
modules, data, and some global declarations. Also, the default
implicit lib is now named `Corelib`.

(lang coq): should be considered deprecated / frozen, and eventually
removed.

There are many cleanups and improvements possible, these will be done
in later commits.

All of the tests have been ported and manually verified to work (a
time consuming process) As of today, all test pass.

Signed-off-by: Emilio Jesus Gallego Arias <e+git@x80.org>