refactor: optimize reading files
Using channels to read a file into a string is rather slow and creates a
mutex + a bunch of useless intermediate buffers.

We instead initialize the buffer once before reading.

This optimization has been used internally for ages with great success.
This is an attempt to make a portable version of it.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>

<!-- ps-id: c3b2757d-cc4b-4e6f-b6fc-0dfdd8ed9eda -->

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>