fix(tui): replace tabs with spaces
Previously dune_tui would raise an exception when coming across control
characters such as `\t` in a message. The error could be observed in
_build/log but it would cause dune to exit with no good reason.

We instead replace any tabs we encounter with two spaces. Note that this
doesn't protect against other control characters, but we will for now
assume that this won't be a common occurance.

Signed-off-by: Ali Caglayan <alizter@gmail.com>